Bring histograms back to the future
Summary:
This commit removes the histogram dashboard's dependencies on old-style
framework code, and removes mentions of histograms from core code like
`backend.ts`.

As usual (e.g., #108), I had to delete some minimal tests in
`backendTests.ts` to do this. Unfortunately, this time there were also
some substantive tests of the `convertBins` function, which I regret
leaving behind. I've included a TODO in the code to add them back in
once our frontend tests work again, which will of course be Soon™.

Note that both the old and new versions of the histogram dashboard have
an unfortunate behavior regarding graphical updates after changing run
selection state. Specifically, suppose that Histogram A and Histogram B
appear side-by-side, and the run for Histogram A is toggled off. Then we
would like Histogram B to seamlessly move into the position of
Histogram A. However, due to Polymer's data binding, the actual action
is that Histogram B is deleted and Histogram A's data bindings are set
to those of the former Histogram B. The result is that the title, tag,
and card color of Histogram A immediately update to those of
Histogram B, but the actual data takes a while to fetch. Furthermore,
this means that _all_ histograms that are affected will re-render their
data. (So if the line goes down to Histogram Z, then removing the run
for Histogram A will cause 25 network requests.)

Essentially, we want [React's keys behavior][1], but it doesn't look
like we can get it. I'm leaving this wontfix for now.

[1]: https://facebook.github.io/react/docs/lists-and-keys.html#keys

Finally, while I'm at it, I fixed all closure compiler warnings in
`vz-histogram-timeseries` and `tf-histogram-dashboard`.

Test Plan:
Tested with the MNIST data combined with the histogram demo data.
Toggled runs, changed histogram modes and time modes, used tag groups,
expanded and collapsed categories. All bazel tests pass, but there
aren't actually any of interest, so that really just means that
everything still builds.

wchargin-branch: refactor-histograms-dashboard